further refactoring

+++ 01 - Code/LIFE/LIFE Services/Agents/GenericAgentArchitectureCommon/Datatypes/Direction.cs
@@ -7,15 +7,15 @@ namespace GenericAgentArchitectureCommon.Datatypes {
   /// </summary>
   public class Direction {
     
-    public float  Pitch { get; private set; }   // Direction (lateral axis).
-    public float  Yaw   { get; private set; }   // Direction (vertical axis).     
+    public double  Pitch { get; private set; }   // Direction (lateral axis).
+    public double Yaw { get; private set; }   // Direction (vertical axis).     
 
 
     /// <summary>
     ///   Set the agent's pitch value [-90° ≤ pitch ≤ 90°].
     /// </summary>
     /// <param name="pitch">New pitch value.</param>
-    public void SetPitch (float pitch) {
+    public void SetPitch(double pitch) {
       if (pitch >  90) pitch =  90;
       if (pitch < -90) pitch = -90;
       Pitch = pitch;
@@ -26,7 +26,7 @@ namespace GenericAgentArchitectureCommon.Datatypes {
     ///   Set the agent's orientation (compass heading, [0° ≤ yaw lt. 360°].
     /// </summary>
     /// <param name="yaw">New heading.</param>
-    public void SetYaw (float yaw) {
+    public void SetYaw(double yaw) {
       yaw %= 360;
       if (yaw < 0) yaw += 360;
       Yaw = yaw;
@@ -40,9 +40,9 @@ namespace GenericAgentArchitectureCommon.Datatypes {
     public Vector GetDirectionalVector() {     
       var pitchRad = DegToRad(Pitch);
       var yawRad   = DegToRad(Yaw);      
-      return new Vector((float) (Math.Cos(pitchRad) * Math.Sin(yawRad)),
-                        (float) (Math.Cos(pitchRad) * Math.Cos(yawRad)),
-                        (float) (Math.Sin(pitchRad))).GetNormalVector();        
+      return new Vector(Math.Cos(pitchRad) * Math.Sin(yawRad),
+                        Math.Cos(pitchRad) * Math.Cos(yawRad),
+                        Math.Sin(pitchRad)).GetNormalVector();        
     }
 
 
@@ -51,18 +51,18 @@ namespace GenericAgentArchitectureCommon.Datatypes {
     /// </summary>
     /// <param name="vector">The vector to set.</param>
     public void SetDirectionalVector(Vector vector) {
-      SetPitch(RadToDeg((float) Math.Asin(vector.Z / vector.GetLength())));
+      SetPitch(RadToDeg(Math.Asin(vector.Z / vector.GetLength())));
       var yaw = Yaw;
 
       // Check 90° and 270° (arctan infinite) first.      
-      if (Math.Abs(vector.Y) <= float.Epsilon) {
+      if (Math.Abs(vector.Y) <= double.Epsilon) {
         if      (vector.X > 0f) yaw = 90f;
         else if (vector.X < 0f) yaw = 270f;
       }
 
       // Arctan argument fine? Calculate heading then.    
       else {
-        yaw = RadToDeg((float) Math.Atan(vector.X / vector.Y));
+        yaw = RadToDeg(Math.Atan(vector.X / vector.Y));
         if (vector.Y < 0f) yaw += 180f;  // Range  90° to 270° correction. 
         if (yaw      < 0f) yaw += 360f;  // Range 270° to 360° correction.        
       }
@@ -75,7 +75,7 @@ namespace GenericAgentArchitectureCommon.Datatypes {
     /// </summary>
     /// <param name="angle">Angle in degree.</param>
     /// <returns>Angle in radians.</returns>
-    public static float DegToRad(float angle) {
+    public static double DegToRad(double angle) {
       return angle*0.0174532925f;
     }
 
@@ -85,7 +85,7 @@ namespace GenericAgentArchitectureCommon.Datatypes {
     /// </summary>
     /// <param name="angle">Angle in radians.</param>
     /// <returns>Angle in degree.</returns>
-    public static float RadToDeg(float angle) {
+    public static double RadToDeg(double angle) {
       return angle*57.295779f;
     }
   }
